Create TTS.py
This is a program that will be integrated in various aspects and courses that will be provided by us once we develop the website. Text to speech program will help majorly the dyslexic students and other neuro divergent students who have problem in reading texts. This is also supposed to strengthen their pronounciation and spelling recognition.